Added Interface to be used by interpreter to parse commands, including a proof-of-example 'move' command.
