+++ src/main/java/interpreter/commands/MoveCommand.java
@@ -1,15 +1,11 @@
-package main.java.utilties.scripting;
+package main.java.interpreter.commands;
 
+import main.java.interpreter.PlayerScriptCommand;
 import main.java.utilties.entities.Player;
 import main.java.utilties.models.Location;
 import main.java.utilties.models.Map;
 
-/**
- * Basic implementation of functions for Scripting Language
- */
-public class PlayerFunctions {
-
-    //TODO Add More scripts based on scripting language design
+public class MoveCommand implements PlayerScriptCommand{
 
     /**
      * Move the Player to location on Map.  If location is occupied, player will not move.
@@ -17,10 +13,10 @@ public class PlayerFunctions {
      * @param location Location to move to
      * @param player Player to move
      */
-    public static void Move(Map map, Location location, Player player){
+    @Override
+    public void performAction(Map map, Location location, Player player) {
         if(!map.isLocationOccupied(location)){
             map.setLocation(player, location);
         }
     }
-
 }
